requirements.txt
discord.py
sqlalchemy
asyncpg
dotenv